refactoring add student, moved search to separate component
