+++ package.json
@@ -10,6 +10,8 @@
     "@mui/lab": "^5.0.0-alpha.49",
     "@mui/material": "^5.0.2",
     "@mui/styled-engine-sc": "^5.0.0",
+    "@mui/styles": "^5.0.1",
+    "@mui/x-data-grid": "^5.0.0-beta.3",
     "axios": "^0.21.1",
     "bcryptjs": "^2.4.3",
     "cookie-session": "^2.0.0-rc.1",
